Authentication

+++ src/App/Controllers/NotificationController.php
@@ -9,34 +9,55 @@ use Symfony\Component\HttpFoundation\Response;
 class NotificationController
 {
     protected $notificationService;
+    protected $server;
+    protected $response;
 
-    public function __construct($service)
+    public function __construct($service, $app)
     {
         $this->notificationService = $service;
+        $this->app = $app;
+        $this->server = $this->app['oauth_server'];
+        $this->response = $this->app['oauth_response'];
     }
 
     public function getOne($id)
     {
-        return new JsonResponse($this->notificationService->getOne($id));
+        if (!$this->checkAuth()) {
+            return $this->server->getResponse();
+        } else {
+            return new JsonResponse($this->notificationService->getOne($id));
+        }
     }
 
     public function getAll()
     {
-        return new JsonResponse($this->notificationService->getAll());
+        if (!$this->checkAuth()) {
+            return $this->server->getResponse();
+        } else {
+            return new JsonResponse($this->notificationService->getAll());
+        }
     }
 
     public function save(Request $request)
     {
-        $email = $this->getDataFromRequest($request);
+        if (!$this->checkAuth()) {
+            return $this->server->getResponse();
+        } else {
+            $email = $this->getDataFromRequest($request);
 
-        $this->notificationService->save($email);
+            $this->notificationService->save($email);
 
-        return new JsonResponse($email);
+            return new JsonResponse($email);
+        }
     }
 
     public function delete($id)
     {
-        return new JsonResponse($this->notificationService->delete($id));
+        if (!$this->checkAuth()) {
+            return $this->server->getResponse();
+        } else {
+            return new JsonResponse($this->notificationService->delete($id));
+        }
     }
 
     public function getDataFromRequest(Request $request)
@@ -49,5 +70,9 @@ class NotificationController
             "channels" => $request->request->get("channels"),
             "status" => "ADDED"
         );
+    }
+
+    private function checkAuth(){
+        return $this->server->verifyResourceRequest($this->app['request_stack']->getCurrentRequest(), $this->response);
     }
 }
